Ensure we know what template has been used
In ci_gen_kustomize_values, we have a growing number of templates. It
might be hard to know what template was used to generated the
02_ci_data.yaml snippet, unless we add a small comment in the templates
directly, pointing to its origin.

This patch adds such comment to the existing files, and also adds a
small "linter" ensuring we match the comment in those templates.