Allow some retries for setup_molecule run
We see some unrelated failures while running the pod, during setup
stage: apparently, the DNS may fail to answer in time, and this crashes
the env setup.

With this patch, we should be in better position, since it would retry
to install the libs in case of failure, with a small delay.

Hopefully, this change will stabilize enough the job so that we may
consider it "voting" in a near future.