Create local PV directories with 0775 permissions
k8s changes permissions of the PV directories when starting eg. an
StatefulSet. However, if the playbook is rerun after starting a
deployment, the group permissions will be reset to read-only and
services can't create new subdirectories.

Changing the group permissions to writeable should be ok for new
directories, as the default group is root until the PV gets mounted and
prevents service errors if the playbook is re-run.