Correct import vs include, apply tags
The 06-deploy-architecture.yml playbook exposes a series of useful tags,
but due to the differences between import and include, some tasks
weren't running when we "select" them with a tag.

There were also issues with import vs include in the play:
- include will dynamically load the content when the play hits that
  task; any tag you set there will be applied only to that "include"
  action, and not reported to the included content.

- import, on the contrary, load all the content during the catalog
  build, and any tag (or `when` condition) set on that import task will
  be applied down to any of the imported content.

In short, when there is no condition on the import, let's use
`import_*`.
If we want to apply the tags down included content, we have to use
`apply` directive.

In order to target precisely the "deploy rhoso" part, new tags were
needed:
- edpm_bootstrap
- edpm_post

With those new tags, in case you want to iterate only on the "generate
CRs" for instance, you can call this:
```Bash
[controller-0]$ ./deploy-architecture.sh --tags edpm \
        --skip-tags edpm_bootstrap,edpm_post \
        -e cifmw_kustomize_deploy_generate_crs_only=true
```
This will target only the necessary tasks to start generating the CRs.