Switch to names instead of ID
With this patch, we allow users to target a stage using a name instead
of a numerical ID.

This should ensure a more stable interface to the users in case we add
new stages in the automation scenario.

The "name" is built using the `build_output` basename, without the
extension.

For instance, if the build_output is "../control-plane.yaml", the stage
name will be `stage_control-plane` (but would also be accessible from
the ID).

Regarding injected hooks: since the run_hook role doesn't filter on
multiple strings (and will not allow it, at least for now), hooks are
still using the ID, and only the ID.