Move away from plain `oc` command in favor of kubernetes.core
In #1562, a fix was imported 1:1 from install_yamls, with command calls.

This lead to some issues with custom environment, where the PATH wasn't
properly set due to the fact, delegate_to and other considerations.

Moving to kubernetes.core collection makes the fix cleaner, while
avoiding any weird environment issues.

We "hardcode" the kubeconfig path because, due to the delegation, we
can't rely on facts.