Enable proper logging
Since we're using zuul pod, we're losing some logs, masked by the
community.general.make module.

Adding the capacity to `tee` in a log file should ensure we keep track
of the potential issues.

This patch modifies all of the test targets in the Makefile to use the
new LOG_DIR parameter.