Prepare for job dependency
In order to save resources and time, we want to make jobs dependent.
Since a patch won't merge if molecule (or pre-commit, or ansible-test)
don't pass, let's not run the full deploy if those smaller, faster tests
fail.

To get that dependency, we leverage the `provides` parameter for zuul
job definition. The next step is to modify the "big jobs" to `requires`
the listed job(s).

For now, we can depend on molecule and pre-commit
- ansible-test is still in Prow