Extend zuul pod usage in ci-framework CI
In order to get out of the Prow CI, we have to port some other jobs to
Zuul pods.

In order to avoid code ducplication, some changes were needed in the
`pre-commit.yml` playbook - for instance, it got renamed with a more
generic name, and it's now getting the `make` target directly from the
job definition.

In addition, two checks that were in pre-commit are now standalone:
- check_zuul_files
- check_k8s_snippets_comment

This ensures we get consistent logging and related artifacts in Zuul, as
well as ensuring we trigger them only when needed (those are small jobs,
but still - let's save resources)