Add creation of the default subnet pools for tenant networks
Default subnet pool is required to run tests for the auto allocate
network functionality. This patch adds creation of such default and
shared pools for IPv4 and IPv6.

This patch is not enough to make auto allocate network tests running
fine as it also require external network to be marked as "default" for
projects and that can't be done using ansible openstack module yet. It
will require patch [1] and then additional change in the os_net_setup
module.

[1] https://review.opendev.org/c/openstack/ansible-collections-openstack/+/917690