Correct import vs include
The `import_role` in 06-deploy-edpm.yml play was leading to an uncaught
issue in architecture driven deployment: with `import`, the content of
the role was directly in the playbook and, while we exit it early via
the ansible.builtin.meta, it was still "too late" for some undefined
variable hit (install_yamls related).

The `include_role` in 06-deploy-architecture.yml play was leading to an
error when we were trying to iterate on `edpm_deploy` tag: the `always`
tag wasn't applied to the subsequent list of tasks, leading to an error
with undefined variable. Switchint to `import` allows to get all the
tasks tagged with `always`.