Use ad-hoc reboot play
This change implements the supported user interface for initiating reboots.
We were using the reboot-os service, but it's expected that this wouldn't work consistently
for day 2 operations. Nothing in the edpm_kernel role creates a reboot_required file and
therefore the default reboot-os service could potentially skip rebooting for changes to
kernel args. This instead implements a custom reboot service that will ensure the reboot is
executed against the defined set of nodes.

Signed-off-by: Brendan Shephard <bshephar@redhat.com>