Allow to deploy architecture against CRC
This patch allows to deploy at least va-hci.yml scenario against CRC.

It's not enough, crafted parameter files are also needed to ensure
proper variables are passed in the right order, but mostly: it works.

It could be used to run architecture validation jobs, at least when VMs
are enough to test things up (for now, HCI, mostly).
When it comes to "baremetal provisioning", sushy may be leveraged to act
as the IPMI/related interface for metal3, allowing to go further down
the path of architecture testing.

This capability isn't "the end" - it's just offering a light way to test
things up.

Documentation will follow to show what specific parameter files are
needed (hence the DRAFT here).