Toggle ansible-lint to offline
In order to avoid seeing the dependencies installed twice, and usually
failing there, let's toggle ansible-lint to offline.

According to the doc:
```
Offline mode disables installation of requirements.yml and schema
refreshing.
```

Let's see what it means for us at this point, and if we can avoid false
flags in the pre-commit job.