Rework imported vars in the reproducer
Until now, we were filtering out anything `cifmw_openshift*` related.
This leads to issues, for instance when we want to pass
`cifmw_openshift_setup_ca_registry_to_add` down in the environment.

In order to ensure we're not filtering too many parameter, this change
modifies a bit how things are handled:
- we allow all of `cifmw_openshift*` parameter
- but we filter out some that may be problematic: kubeconfig (due to
  potential user HOME mismatch), and openshift_token.

We also take the opportunity to reject the pull-secret and ci-token
keys, to ensure they don't leak in the env (they are mostly useless on
controller-0).