Generate all MACs beforehand
We need to generate all the MAC addresses before deploying anything:
this will allow us to stop relying on libvirt to get the IP in further
steps.

The goal here is to generate the "interfaces-info.yml" content before
starting to deploy any VM, and re-use the generated content when we
associate network interfaces to the VMs.

The generator takes care to not override existing data if we're
re-running a deployment, but only adds new content (for instance if you
add new nodes, or new network intefaces).

There's also a list of TODO for the future work on that networking
change.