Fix(?) CRC DNS
In CRC, we can't really configure existing interfaces - they get
overridden by some other process, meaning we can't ensure it won't get
any public DNS.

In order to ensure we use only the local dnsmasq , the best way is
apparently to unmanage the /etc/resolv.conf from within NetworkManager,
and hardcode the content.
Even with that, the comment seems to be removed and replaced by
`# Generated by NetworkManager` - which is... weird. Since it shouldn't
do anything in there...

We use the CRC "public" IPv4 - apparently 127.0.0.1 doesn't work, at
least from the openshift-dns/dns-default pod (throws connections
refused).

In any cases, this change seems to work (for now), ensuring DNS are
properly working.

We might want to amend the DNS configuration on the controller-0 to use
the openshift-dns/dns-default service instead - but that's for a
follow-up.