[test-operator] Make ntp_extra_image more generic
The cifmw_test_operator_tempest_ntp_extra_image was designed with
neutron tempest plugin in mind. It only allows specifying an image that
should be downloaded but does not allow to define details like flavor
for this image or ID that should be assigned to the given image.

This patch removes the ntp_extra_image parameter and introduces a new
cifmw_test_operator_tempest_extra_images parameter that exposes the
extraImages parameter from Tempest CR that allows to specify the more
granular details.