Make start VM and Manage (inventory + jumper) optional
It does not make sense to add all VMs created in a lab environment
to the ci-framework inventory.

Also not starting VMs can be useful even for VMs that don't use a
blank image.

Add two fields to the "vms" definition: `start` and `manage`.
A VM that is not started won't be managed, a VM can be started but
not managed.