[kustomize_deploy] Combine multiple user kustomizations as one
In automation it may happen that the jobs have already defined
a cifmw_architecture_user_kustomize but the user wants to add
his own kutomization or just concat multiple ones as part
of his job parenting strategy. With the previous approach
only one cifmw_architecture_user_kustomize may survive, forcing
the user to be aware of the already defined kustomization
and making him duplicate all the content.
With this approach we can consume multiple kustomization
dicts that will be combined all together before consuming them
as a single one.