[env_op_images] Avoid using install_yamls defaults
Since this role is consumed from the logs playbook, that is
standalone and that should work perfectly fine without install_yamls
just remove the hard dependency it has with the
cifmw_install_yamls_defaults variable.
To do so, it now queries all the namespaces searching the CSV and
it uses the CSV namespace afterwards.