Add review ready label automation
This patch adds a GitHub workflow that runs on each status update and evaluates if the PR is "Ready for Review"

The evaluation steps are:
- There are no failed CI jobs
- There are no Pending CI jobs excluding Tide
- The PR isn't in draft state
- The commit isn't found in more than one PR

If the evaluation passes the review_ready label is applied, if they fail the review_ready is removed.

The end result after this patch is maintainer can monitor the "Ready for Review" label and be confident if a PR has that label, it's actually really for review.