libvirt_manager - enable sushy_emulator
Add a new boolean option:
 `cifmw_libvirt_manager_enable_sushy_emulator` (defaults to: false).

When enabled - include the sushy_emulator role in deploy_layout.yml
to set-up virtual RedFish BMC.

Also change hos sushy_emulator role builds the libvirt_uri, instead
of a lookup from hostvars require the hostname to be defined in
a variable when calling the role. This allows use of the role even
if the libvirt hypervisor is not managed by ansible.