Allow passing base64 encoded patches to ci_gen_kustomize
Some content passed to the ci_gen_kustomize role can
be interpreted by Ansible as a jinja template
(edpm_network_config_template). That content, when ingested by the
framework, can lead to a crash as Ansible may try to use it as a
template. Mechanisms like raw/endraw or !unsafe don't work, as those
tags are removed in the first write of the variable to a file (and the
framework performs one write of those variables before edpm-deploy).
To allow easily passing those kind of variables this mechanism can be
handy, as the content is safely encoded as a plain string before getting
applied.