Mark created external network as default for projects
To run "auto_allocate_network" tests it is required to have external
network which is marked as "default" external network for the projects.

After patch [1] was merged, it can be now done using Openstack ansible
collection directly and this patch adds setting of such flag in the
os_net_setup module.

Related: #OSPRH-6541