[edpm_build_images] Run buildah command as a root
In Downstream, we perform podman/buildah login as a root to pull
base containers.

Here we are building containers as a normal user. the buildah commands are
failing with base containers not found.

Running buildah command as a root fixes the issue.
Since we are building the images as root, we also need to push
the images as root.

Signed-off-by: Chandan Kumar <raukadah@gmail.com>