Switch away from `lookup()`
Since we want to run some validations against `architecture` repository,
we end calling the role against a remote node instead of localhost.

The usage of `lookup()` breaks the loop, especially on Zuul: the
`lookup()` is launched onto the zuul-executor node, where we don't
manage anything, and may lead to multiple issues such as:
- collection not present (it seems to be present, but still)
- kustomize being missing from the path (seems to be present for now)

Calling plain `command` ensures we're running the build onto an
environment we 100% manage and know.

This patch also fixes molecule tests on kustomize_deploy.

Signed-off-by: Roberto Alfieri <ralfieri@redhat.com>