[networking_mapper] Switch to best effort strategy
With the introduction of IPv6 the mapper now needs to perform
partial maps with hosts in the inventory that are not yet
reachable, so it needs to be able to perform the mapping without
a full set of ansible facts.

The strategy that the PR follows is the next one:
- Despite the type of mapping, full or partial, do a best effort
approach without failing if a fact is missing.
- In partial maps, allow the mapper to take the interfaces-info
- Avoid refreshing networking facts for those hosts that has
not ansible_host fact present, assuming those are not reachable.