Refine a bit dnsmasq role
An assertion block was missing a condition, leading to issue when we
wanted to just remove a network from the configuration.

In the network configuration template, the test to know more about ipv4
and ipv6 ranges was a bit too loose, leading to issues under certain
conditions.

We faced the same issue in the host management, leading to potentially
wrong DHCP configuration. In order to ease debugging, we now generate a
fact, display its content, and use that content to inject the host data
in the configuration.

During a re-run, the service is already stopped/disabled, and the unit
file is removed. Ansible will then fail, since systemd can't find the
service to stop.