[ci_gen_kustomize] Remove the dedicated ctlplaneInterface var
As now the cifmw_baremetal_hosts is used outside of the scope of
real BM in NFV it can happen that we have preprovisioned
nodes filled in that variable. In those circumstances we won't
need to fill the baremetal set with that variable.
To avoid the framework to unreliably decide when that variable
is needed let the user pass it if needed like the rest of the
EDPM variable we are consuming.