Allow to patch devscripts config overrides
The Framework (or a job) might need to inject specific configurations in
devscripts.

This pach proposes an interface to inject as many configuration snippets
as needed - in the exact same way other roles expose such feature.

As the molecule test has been updated, it now hits a specific path in
the deployment that wasn't tested before. In molecule, it leads to a
failure, because `firewalld` isn't present.
This patch also ensures this package is installed.