Allow to patch devscripts config overrides
The Framework (or a job) might need to inject specific configurations in
devscripts.

This pach proposes an interface to inject as many configuration snippets
as needed - in the exact same way other roles expose such feature.

As the molecule test has been updated, it now hits a specific path in
the deployment that wasn't tested before. In molecule, it leads to a
failure, because `firewalld` isn't present.
This patch also ensures this package is installed.

We also expose the "build devscripts config" as a dedicated tasks file,
allowing to get it "on-demand", and avoid code duplication:
it might happen someone wants to just know if there's an OCP cluster or
not, so patching the 110_check_ocp.yml to call the "config builder" if
needed allows to call it directly.

Added fix:
this patch also fix a missed renamed from `_layout` to
`_cifmw_libvirt_manager_layout`.