Refine a bit validate-architecture related make targets
This patch refines a bit the targets involved in the
"architecture_test". It now creates a new volume to store your generated
snippets, allowing to review them once the container is removed.
It also allows to pass any arbitrary `ansible-playbook` options down to
the test run, so that you can point to additional environment files,
pass extra parameters and so on.