Avoid failures on cleaning
While working on integrating dnsmasq role in the reproducer, a random
issue, related to the facts, might happen.

It seems to be caused by the way ansible manages the handler - the
`ansible_facts.packages` isn't available anymore, and the condition
triggering the libvirt cleanup fails. While it was already done.

This condition arrises when dnsmasq role is in the game, and the handler
kicks at the end of the playbook.

In the same fashion, the handler might be triggered when the service is
already removed, leading to another crash of the cleanup playbook. The
added conditions ensure the handler will succeed even if the service is
already removed, but would still fail if there's a real issue with the
service configuration.