Generate all MAC addresses beforehand
Pre-generating all the MAC, for all the instances, on all networks,
ensures we can stop relying on libvirt to learn about that data.

This also changes a bit how we gather logs from the reproducer molecule.
It's needed here in order to properly debug failures occuring only
upstream (not reproduceable on a local lab).