Fix sushy_emulator integration in reproducer
A series of assumptions were made and lead to some issues with the way
the reproducer works.

Mostly, it can happen a hypervisor doesn't have the "zuul" user (for
example a home lab), leading to issues when facts such as
`ansible_user_id`, `ansible_user_dir` and related are used in conjuction
with `delegate_to`.

This patch allows the reproducer to properly works with the
sushy_emulator integration, and fixes some of those assumptions.

It also ensure the molecule tests for reproducer role is triggered uppon
podman and sushy_emulator role modifications.