Fix all IP addresses for all known VMs on all known networks
In order to get there, we have to deploy dnsmasq service on the
reproducer, while ensuring it's not already running via NetworkManager
(this happens when CRC has been setup on the hypervisor).

Then, we generate specific network configuration in dnsmasq for each
managed interface - we can also NOT use dnsmasq for selected interfaces,
such as `ocppr`, used by devscripts to deploy OCP cluster.

To get the fixed IPs, we're using the networking_mapper, ensuring we're
getting IPs in the selected networks, without any collision, and just
inject a generated mapping IP, hostname and MAC in dnsmasq
configuration.

Note that dnsmasq is initialized in
libvirt_manager/generate_networking_data since that's the place we get
all the virtual network data, among them the interfaces to listen to.