nit: Ensure sushy is installed after base packages
Base packages are installed using `async` to allow generating most of
not all the relevant content instead of waiting.

This leads to a race condition, where sushy_emulator role wants to
install dependency packages and fails, because the RPM database is still
locked with the dependency async run.

This patch moves the sushy bootstrap and installation to a later step,
ensuring we're not coliding with the async task.