Ensure no default network is around
the `default` network is usually creating issues with the range overlap
for osp_trunk managed net.

`crc` network doesn't create issue *now* but will be an issue once we
get our own dnsmasq integrated: there might be some service collision
with the DNS and DHCP services listening on that network.