Add ssh_jumper role
Currently the ssh jumper is configuerd by tasks in the
libvirt_manager role. Adding jumper configuration may be
useful for other roles that create VMs outside of the ones
defined in the "layout" deployed by the libvirt_manager
tasks.

This change adds a new role `ssh_jumper` - and updated the tasks
in libvirt_manager role to use this new role when creating the
SSH config.