Re-add the host IPs in controller-0 .ssh/config, revert hostname
There's some magical things between .ssh/config and ansible that were
broken for HCI, when time comes to deploy ceph using the ceph.yml
playbook.

In parallel, this patch also reverts the hostname of the VMs to its
former "short" format (compute-0, etc) instead of a more FQDN-like
one (compute-1.hypervisor-1, etc).
That change was mostly needed for sushy since it consumes, by default,
the `ansible_host`.
Injecting that value in the /etc/hosts makes it easier to consume.