dnsmasq - use dhcp-hostdir
The restart done by the handler happens only at the very
end of the play. By using dhcp-hostsdir we get the benefit
of dnsmasq reading any new or changed file in the hosts dir,
automatically applying the new dhcp host config on the fly.

Since adding a host is loaded on-the-fly - we don't need to
notify the handler when adding hosts.

Removing hosts however, reqires s SIGHUP - so for that case
notify the handler.

Also change the computation of _host_entry - it should include
both ipv4 and ipv6 address if both are configured.