polarion: update the usage of cifmw_polarion_update_testcases
cifmw_polarion_update_testcases is named after a parameter of a tool
and it doesn't match its behavior: it is about adding missing test cases
when they don't exist. But the "normal" workflow is to pre-populate
the test cases and only update (for real) their results,
so that variable should almost never be enabled.

Also, remove the check about the variable being enabled,
as its only usage is guarded by a default() filter.