Add retries/delays when applying LVMS manifests
In the `ci_lvms_storage` role, the kubernetes.core.k8s
module will occasionally attempt to apply a manifest
before an end point is available for a service which was
created in the previous task as reported in OSPRH-7460.

Add retries and delays to all calls to this module within
the role so that it retries if the module fails.

Closes: https://issues.redhat.com/browse/OSPRH-7460

Signed-off-by: John Fulton <fulton@redhat.com>