Move task for reading parent file to task_file for reuse
Move the reading parent scenario to a task, so it can be used in the
reproducer and deploy-edpm playbooks. This change also introdcues
inhertance for the hci VA scenarios. The parent scenario is moved to
before the task including the common va, since otherwise we might be
missign the cifmw_architecture_scenario vars.