Change iscsid login timeout
This patch changes the time to fail an iSCSI login request in our
deployments.

The default is 8 retries and 15 seconds for each one, so 120 seconds in
total.

This default timeout is too large for our purposes, because it's greater
than the default HAProxy or Apache timeouts.  This will create problems
creating an image when Glance is using Cinder as a backend and one of
the paths to the storage is down.

This patch changes the OpenShift cluster default to 3 retries and 5
seconds each (15 seconds in total), which is convenient for testing, as
any healthy deployment and backend should be able to login to the
backend in that amount of time, and if there is a broken path it will
not take 2 minutes to give up, just around 15 seconds.

Jira: https://issues.redhat.com/browse/OSPRH-7393
Jira: https://issues.redhat.com/browse/OSPRH-7415