uni01alpha - update network-values values.j2 for ironic
In PR: https://github.com/openstack-k8s-operators/architecture/pull/211
the DT is updated to configure Ironic so that it can be used beyond
running API tempest test.

It adds a network bridge "ironic", and this is attached to the ironic
service pods, OVN. Also Nova is configured to run a compute pod with
the ironic driver.