Introduce cifmw_external_dns role to configure RGW with TLS
Ceph RGW is hosted on EDPM nodes and used by OpenStack pods.
In order for Ceph RGW to use HTTPS we need to create a
TLS certificate for it which uses OpenStack's public root
CA. We also need OpenStack pods to resolve the hostnames
of the RGW endpoints so there are no certificate errors.

The cifmw_external_dns role introduced in this PR solves
the DNS and Certificate requirements for RGW and the same
role may be used to solve the same problem by other services
which are hosted externally (outside of k8s).

Jira: https://issues.redhat.com/browse/OSPRH-7245

Signed-off-by: John Fulton <fulton@redhat.com>