Converge namespace for hooks
When running architecture driven deployment, install_yamls isn't
initialized anymore since it's not used.

This patch allows to converge how we pass the namespace down to the
hooks, by checking if we have the install_yamls data and defaulting to a
"not too bad" default value.

In case a user wants to override that namespace, they can use the
extra_vars parameter - since that content comes after our default
setting, it will take over.