Use host_ip in _sushy_url
The `ansible_host` contains a name only available in the ssh client
configuration. This results in connection URI with a name that cannot be
resolved:
 connection: redfish-virtualmedia+http://controller-0.hypervisor-1:8000/[...]

Instead of using ansible host, use the host_ip which contains the IP address.

This is an intermidiate fix, once we have the changes to use DNS more
extensively we should change this to use the FQDN of the controller - or
possibly add a CNAME in DNS with `sushy-emulator.$DOMAIN`.