Can't use ULA addresses in NAT64
Ran into an issue where routing on the OCP cluster gets confused by the
use of ULA addresses on the machine network.

OCP uses ULA addresses for internal networks, OVN kubernetes for
example. When using ULA addressing on the machine network there is race
when building the routing tables - and in some cases the hosts end up
using the wrong source address in outgoing traffic.

To fix this, switch to GUA addresses.