Remove change_message from zuul parameters
After a recent Zuul update, zuul was putting the first comment of the PR
as the change_message in the zuul var. If that message contained yaml,
it would get parsed and in all likelihood cause an error. This change
removes the change_message field from the zuul var to allow yaml to be
used in the PR comment.