[tcib]Use inner ansible vars to expose dlrn md5 hash
Currently we use to slurp the content of dlrn.md5 file via run
playbook running from zuul executor. Most of the time, It does not
exposes the correct dlrn md5 file hash. In order to fix that,

We are moving log collection from zuul post run to tcib playbook to
create the ansible var file and use the same to read the dlrn md5 hash
file.

Signed-off-by: Chandan Kumar <raukadah@gmail.com>