Add a new extra variable for holding the update's parameters
Make sure that, when we use downstream repo setup we don't hardcode
cifmw_repo_setup_promotion.

To that end, and to add more flexibility to the variables that need to
be overwritten during update, take the key/value of the
`cifmw_update_extras` hash. It should be defined in a job and would
look like this:

```
cifmw_update_extras:
 cifmw_repo_setup_promotion: promotion
 other_var: value
 ...
```

That enables the job definition to pass all variables needed for the
update job with having new variable name for each override.

It also solves the problem of the hardcoded
`cifmw_repo_setup_promotion` as now the job has total control over the
variables passed.