Ensure virtqemud-sock is present before cleaning
While the usual usage is "deploy, then clean", jobs in CI might start
with a "clean, then deploy".

In such case, it might happen all of the dependencies are installed
during the node bootstrap, but libvirt services aren't switched to the
modular daemons layout, leading to a crash with the current
implementation.

This patch adds a sanity check on the virtqemud-sock file availability,
to ensure everything's fine.

Molecule job has also been modified to "clean" before "deploy" so that
we should hopefully catch issues earlier.

Signed-off-by: Roberto Alfieri <ralfieri@redhat.com>