Ensure CRC is using our new dnsmasq
Since #1777, a special dnsmasq instance is created and managed by the
ci-framework.
While most of the testing was done with OCP cluster in mind, we have to
ensure CRC is able to properly consume that new service.

This allows to remove all of the hacks we had in CRC node related to
DNS, making everything really easier for everyone.

Since we now have proper DHCP on the various networks, with fixed IPs
and all the rest, we don't need to connect to the CRC instance anymore:
it gets its second NIC with appropriate configuration since the boot
time, and the DNS service is properly answering.

Proofs of work:
```
[zuul@controller-0 ~]$ oc rsh dns-default-68hj2 dig +short api.crc.testing
Defaulted container "dns" out of: dns, kube-rbac-proxy
192.168.122.10
[zuul@controller-0 ~]$ oc rsh dns-default-68hj2 dig +short foo.apps.crc.testing
Defaulted container "dns" out of: dns, kube-rbac-proxy
[zuul@controller-0 ~]$ oc rsh dns-default-68hj2 dig +short glab.testing
Defaulted container "dns" out of: dns, kube-rbac-proxy
[zuul@controller-0 ~]$ oc rsh dns-default-68hj2 dig +short fff.api-int.crc.testing
Defaulted container "dns" out of: dns, kube-rbac-proxy
192.168.122.10
[zuul@controller-0 ~]$ oc rsh dns-default-68hj2 dig +short google.com
Defaulted container "dns" out of: dns, kube-rbac-proxy
142.250.203.110
```